Generate package set on-the-fly
Instead of a library containing a manually written build-depends
corresponding to a stackage snapshot, we now have an executable that
queries stackage directly, and then uses the response to generate
the desired cabal file. The executable then builds that project.

The executable also includes the ability to split the package set into
smaller groups, where each group is built sequentially. This allows
for scenarios where building the entire set at once is not feasible,
at the possible cost of some performance.